Resolve build error - move types outside of pages folder
