+++ types/BlogInterfaces.ts
